refractor(data-source):moving-the-orientedURL-generation-to-a-shared-file
